+++ src/config.rs
@@ -6,6 +6,8 @@ use std::{collections::HashMap, fs};
 use std::{convert::TryFrom, io::ErrorKind};
 use walkdir::WalkDir;
 
+use crate::Vcs;
+
 pub const CONFIG_FILE_NAME: &str = "cargo-generate.toml";
 
 #[derive(Deserialize, Debug, PartialEq, Default, Clone)]
@@ -28,6 +30,8 @@ pub struct TemplateConfig {
     pub include: Option<Vec<String>>,
     pub exclude: Option<Vec<String>>,
     pub ignore: Option<Vec<String>>,
+    pub vcs: Option<Vcs>,
+    pub init: Option<bool>,
 }
 
 #[derive(Deserialize, Debug, PartialEq, Clone)]
@@ -51,21 +55,19 @@ impl TryFrom<String> for Config {
 }
 
 impl Config {
-    pub(crate) fn from_path<P>(path: &Option<P>) -> Result<Option<Self>>
+    pub(crate) fn from_path<P>(path: &Option<P>) -> Result<Self>
     where
         P: AsRef<Path>,
     {
         match path {
             Some(path) => match fs::read_to_string(path) {
-                Ok(contents) => Self::try_from(contents)
-                    .map(Option::from)
-                    .map_err(|e| e.into()),
+                Ok(contents) => Self::try_from(contents).map_err(|e| e.into()),
                 Err(e) => match e.kind() {
-                    ErrorKind::NotFound => Ok(None),
+                    ErrorKind::NotFound => Ok(Self::default()),
                     _ => anyhow::bail!(e),
                 },
             },
-            None => Ok(None),
+            None => Ok(Self::default()),
         }
     }
 
@@ -174,7 +176,7 @@ mod tests {
         )
         .unwrap();
 
-        let config = Config::from_path(&Some(config_path)).unwrap().unwrap();
+        let config = Config::from_path(&Some(config_path)).unwrap();
 
         assert_eq!(
             config.template,
@@ -183,6 +185,8 @@ mod tests {
                 include: Some(vec!["Cargo.toml".into()]),
                 exclude: None,
                 ignore: None,
+                vcs: None,
+                init: None,
             })
         );
         assert!(config.placeholders.is_some());