+++ tests/integration/basics.rs
@@ -1,4 +1,5 @@
 use git2::Repository;
+use indoc::indoc;
 use predicates::prelude::*;
 
 use crate::helpers::project::binary;
@@ -1277,6 +1278,46 @@ version = "0.1.0"
 }
 
 #[test]
+fn vsc_none_can_be_specified_in_the_template() {
+    // Build and commit on branch named 'main'
+    let template = tmp_dir()
+        .file(
+            "Cargo.toml",
+            r#"[package]
+name = "{{project-name}}"
+description = "A wonderful project"
+version = "0.1.0"
+"#,
+        )
+        .file(
+            "cargo-generate.toml",
+            r#"[template]
+vcs = "None"
+"#,
+        )
+        .init_git()
+        .build();
+
+    let dir = tmp_dir().build();
+
+    binary()
+        .arg("generate")
+        .arg("--git")
+        .arg(template.path())
+        .arg("--name")
+        .arg("foobar-project")
+        .current_dir(&dir.path())
+        .assert()
+        .success()
+        .stdout(predicates::str::contains("Done!").from_utf8());
+
+    assert!(dir
+        .read("foobar-project/Cargo.toml")
+        .contains("foobar-project"));
+    assert!(Repository::open(dir.path().join("foobar-project")).is_err());
+}
+
+#[test]
 fn it_provides_crate_type_lib() {
     // Build and commit on branch named 'main'
     let template = tmp_dir()
@@ -1433,4 +1474,46 @@ fn error_message_for_invalid_repo_or_user() {
             predicates::str::contains(r#"Error: Please check if the Git user / repository exists"#)
                 .from_utf8(),
         );
+}
+
+#[test]
+fn a_template_can_specify_to_be_generated_into_cwd() -> anyhow::Result<()> {
+    let template = tmp_dir()
+        .file(
+            "Cargo.toml",
+            indoc! {r#"
+                [package]
+                name = "{{project-name}}"
+                description = "A wonderful project"
+                version = "0.1.0"
+                "#},
+        )
+        .file(
+            "cargo-generate.toml",
+            indoc! {r#"
+                [template]
+                init = true
+                "#},
+        )
+        .init_git()
+        .build();
+
+    let dir = tmp_dir().build();
+
+    binary()
+        .arg("gen")
+        .arg("--git")
+        .arg(template.path())
+        .arg("-n")
+        .arg("foobar-project")
+        .arg("--branch")
+        .arg("main")
+        .current_dir(&dir.path())
+        .assert()
+        .success()
+        .stdout(predicates::str::contains("Done!").from_utf8());
+
+    assert!(dir.exists("Cargo.toml"));
+    assert!(!dir.path().join(".git").exists());
+    Ok(())
 }